Move common registry, transform, and constants to _shared
