+++ workflows/diagnostics/fv3net/diagnostics/_shared/registry.py
