+++ workflows/diagnostics/fv3net/diagnostics/prognostic_run/views/static_report.py
@@ -21,7 +21,12 @@ from .matplotlib import (
     plot_cubed_sphere_map,
     plot_histogram,
 )
-from ..constants import PERCENTILES, PRECIP_RATE, TOP_LEVEL_METRICS, MovieUrls
+from fv3net.diagnostics._shared.constants import (
+    PERCENTILES,
+    PRECIP_RATE,
+    TOP_LEVEL_METRICS,
+    MovieUrls,
+)
 
 import logging
