+++ workflows/diagnostics/tests/prognostic/test_savediags.py
@@ -1,5 +1,5 @@
 from itertools import chain
-from fv3net.diagnostics.prognostic_run.constants import DiagArg
+from fv3net.diagnostics._shared.constants import DiagArg
 
 import fv3net.diagnostics.prognostic_run.compute as savediags
 from fv3net.diagnostics.prognostic_run.load_run_data import (
